Consolidate OtlpResource tests
